Prevent Harvester from trying to run on Linux (BL-8475)
At the moment, phash computation results can differ between Windows and
Linux.  This would have to be dealt with somehow before we could run
Harvester on a Linux machine instead of a Windows machine.

(This is currently moot because Harvester won't even build on Linux, but
this change should give pause to moving in that direction.)